requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
appdirs==1.4.4
Appium-Python-Client==1.0.2
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
enum34==1.1.10
html-testRunner==1.2.1
idna==2.10
isort==5.5.2
Jinja2==2.11.2
jmespath==0.9.5
lazy-object-proxy==1.4.3
loguru==0.4.1
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
namedlist==1.8
packaging==20.4
pathspec==0.8.0
Pillow==8.0.0
pluggy==0.13.1
py==1.9.0
pydantic==1.6.1
pydash==4.8.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-xdist==2.2.0
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
retry==0.9.2
selenium==3.141.0
sentry-sdk==0.14.4
six==1.15.0
some-package==0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.24.3
verify==1.1.1
wcwidth==0.2.5
win32-setctime==1.0.2
wrapt==1.12.1
